Rename imports as well
